Fixed most compiler warnings

+++ GameEngine/src/renderer/Skeleton.cpp
@@ -7,7 +7,7 @@ Skeleton::Skeleton()
 
 Skeleton::Skeleton(const std::string& name, const std::vector<Joint>& joints, const std::map<std::string, Animation>& animations)
 {
-	NumJoints = joints.size();
+	NumJoints = (unsigned int)joints.size();
 
 	m_Name = name;
 	m_Joints = joints;
